my favorite esp. 6
For me, the best episode of Squid Game is Episode 6: “Gganbu.” It’s the most emotional part of the series because it shows how trust can turn into heartbreak. When the players found out they had to compete against their partners, it shocked me. The scene between Gi-hun and Il-nam really stood out—Gi-hun tried to trick him, but Il-nam knew all along. It showed how desperation changes people. This episode is the best for me because it wasn’t just about survival, it was about humanity and sacrifice.